# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "0.1",
        "titel": "API Sentinelle"
    },
    "servers": [
        {
            "url": "http://api.sentinelle.com/v1",
            "description": "Api de Sentinelle"
        }
    ],
    "paths": {
        "/ecoles": {
            "get": {
                "tags": [
                    "Ecoles"
                ],
                "operationId": "App\\Controller\\EcoleController::index",
                "responses": {
                    "200": {
                        "description": "Liste des ecoles inscritent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ecole"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    "bearer"
                ]
            },
            "post": {
                "tags": [
                    "Ecoles"
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/EcoleUpdate"
                },
                "responses": {
                    "200": {
                        "description": "Detail d'une ecole en particulier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ecole"
                                }
                            }
                        }
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        },
        "/ecoles/{id}": {
            "get": {
                "tags": [
                    "Ecoles"
                ],
                "operationId": "App\\Controller\\EcoleController::show",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail d'une ecole en particulier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EcoleShow"
                                }
                            }
                        }
                    }
                },
                "security": [
                    "bearer"
                ]
            },
            "put": {
                "tags": [
                    "Ecoles"
                ],
                "operationId": "App\\Controller\\EcoleController::edit",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/EcoleUpdate"
                },
                "responses": {
                    "200": {
                        "description": "Detail d'une ecole en particulier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ecole"
                                }
                            }
                        }
                    }
                },
                "security": [
                    "bearer"
                ]
            },
            "delete": {
                "tags": [
                    "Ecoles"
                ],
                "operationId": "App\\Controller\\EcoleController::delete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "security": [
                    "bearer"
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Securities"
                ],
                "operationId": "App\\Controller\\SecurityController::register",
                "requestBody": {
                    "request": "Register",
                    "description": "Corp de la requete",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "telephone": {
                                        "required": true,
                                        "type": "string",
                                        "example": "+24104609874"
                                    },
                                    "passeword": {
                                        "required": true,
                                        "type": "string",
                                        "example": "emileA15ans"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Security"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    "bearer"
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Ecole": {
                "description": "Ecole",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Lycée Privée Saint Esprit"
                    },
                    "devise": {
                        "type": "string",
                        "example": "Union - Travail - Justice"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EcoleShow": {
                "description": "Ecole",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Ecole"
                    }
                ]
            },
            "Security": {
                "description": "Inscription a l'API Sentinelle",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+241 74 60 98 74"
                    },
                    "passeword": {
                        "type": "string",
                        "example": "couCou45"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "NotFound": {
                "description": "La resource n'existe pas",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "L'article n'existe pas"
                                },
                                "status": {
                                    "type": "integer",
                                    "example": 404
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID de la resource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "requestBodies": {
            "EcoleUpdate": {
                "description": "Valeurs du formulaire",
                "required": [
                    "nom"
                ],
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "nom": {
                                    "type": "string",
                                    "example": "Lycée Leon Mba"
                                },
                                "devise": {
                                    "type": "string",
                                    "example": "Union - Travaile - Justice"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}